rename testing_helpers to _testing.memory_threshold
